Рефакторит структуру проекта
