+++ jsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+      "baseUrl": "./",
+      "paths": {
+          "#*": ["./src/*"]
+      }
+  }
+}
