+++ src/eleventy-config/filters.js
@@ -1,0 +1,56 @@
+const hyphenLibRu = require('hyphen/ru');
+
+module.exports = function(eleventyConfig) {
+    eleventyConfig.addFilter('limit', (array, limit) => {
+        return array.slice(0, limit);
+    });
+
+    eleventyConfig.addFilter('addHyphens', (content, maxLength = 0) => {
+        if (!content || content.length <= maxLength) {
+            return content;
+        }
+        let contentWithHyps = hyphenLibRu.hyphenateSync(content);
+        return contentWithHyps;
+    });
+
+    eleventyConfig.addFilter('fixLinks', (content) => {
+        const reg = /(src="[^(https://)])|(src="\/)|(href="[^(https://)])|(href="\/)/g;
+        const prefix = `https://web-standards.ru` + content.url;
+        return content.templateContent.replace(reg, (match) => {
+            if (match === `src="/` || match === `href="/`) {
+                match = match.slice(0, -1);
+                return match + prefix;
+            } else {
+                return match.slice(0, -1) + prefix + match.slice(-1);
+            }
+        });
+    });
+
+    // Даты
+
+    eleventyConfig.addFilter('ruDate', (value) => {
+        return value.toLocaleString('ru', {
+            year: 'numeric',
+            month: 'long',
+            day: 'numeric',
+        }).replace(' г.', '');
+    });
+
+    eleventyConfig.addFilter('shortDate', (value) => {
+        return value.toLocaleString('ru', {
+            month: 'short',
+            day: 'numeric',
+        }).replace('.', '');
+    });
+
+    eleventyConfig.addFilter('isoDate', (value) => {
+        return value.toISOString();
+    });
+
+    eleventyConfig.addFilter('markdown', (value) => {
+        let markdown = require('markdown-it')({
+            html: true,
+        });
+        return markdown.render(value);
+    });
+};
